pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "draftkid"                     # repo name (can be anything)
version = "0.1.0"
description = "nfl_data_py-first fantasy draft builder"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=1.5,<2.0",
  "numpy>=1.24",
  "PyYAML>=6.0",
  "typer>=0.12.0",
  "rich>=13.0",
  "nfl_data_py>=0.3.3",
  "scikit-learn>=1.2,<1.4",
]

[tool.setuptools]
package-dir = {"" = "src"}           # use the /src layout
packages = ["draftkit"]              # this must match src/draftkit

[project.scripts]
draftkit = "draftkit.cli:app"        # optional: enables `draftkit ...` shell cmd

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--cov=src/draftkit",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=75"
]